+++ src/readFile.h
@@ -14,7 +14,7 @@ FileBuffer* openFile (const char* filename, uint32_t bufferSize);
 void closeFile (FileBuffer* buffer);
 uint32_t addToPosition(FileBuffer* buffer, uint32_t amount);
 uint32_t getPosition(FileBuffer* buffer);
-uint8_t getValue (FileBuffer* buffer, uint32_t offset);
+uint8_t getValue (FileBuffer* buffer, int64_t offset);
 int isEndOfFile(FileBuffer* buffer);
 int isEndOfFileAmount (FileBuffer* buffer, uint32_t amount);
